ticket-Master/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
Flask-Mail==0.9.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
requests==2.31.0
PyJWT==2.8.0
bcrypt==4.0.1
Pillow==10.0.0
python-dateutil==2.8.2
pytest==7.4.0
pytest-cov==4.1.0
pytest-flask==1.2.0
gunicorn==21.2.0
paypalrestsdk==1.7.1
